pyproject.toml
[tool.poetry]
name = "pythonproject"
version = "0.1.0"
description = ""
authors = ["Ilia Kaziamov <kaziamov@outlook.com>"]
#packages = [
#    { include = "expenses_app" },
#]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.2"
sqlalchemy = "^2.0.21"
asyncpg = "^0.28.0"
python-dotenv = "^1.0.0"
alembic = "^1.12.0"
psycopg2-binary = "^2.9.8"
starlette-admin = "^0.11.2"
aiogram = "^3.1.1"
uvicorn = "^0.23.2"
cachetools = "^5.3.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
flake8 = "^6.1.0"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
